pyproject.toml
[project]
name = "student-picker-tkinter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
]
